requirements.txt
aenum==3.1.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
colorama==0.4.4
commonregex==1.5.4
ecdsa==0.17.0
flatdict==4.0.1
idna==3.2
loguru==0.5.3
multidict==5.1.0
okta==2.0.0
pyasn1==0.4.8
pycryptodome==3.10.1
pydash==5.0.2
python-dotenv==0.19.0
python-jose==3.3.0
PyYAML==5.4.1
rsa==4.7.2
six==1.16.0
slack-bolt==1.7.0
slack-sdk==3.8.0
typing-extensions==3.10.0.0
win32-setctime==1.0.3
xmltodict==0.12.0
yarl==1.6.3
